HappyTools v0.0.1c b170622a
This build fixes two small bugs (#2 and #3). The #2 bug was fixed by
taking the entire time window for analytes that do not have either a
local maxima or minima within that time window. The #3 bug was fixed by
string formatting the number, using a user setting that specifies the
number of decimal values (variable is `decimalNumbers` specified in
`functions.py`).