HappyTools v0.0.2 b170714a
This build implements a calibration check, meaning that the program
calculates the S/N for each potential calibrant peak and only uses the
peaks that pass the specified S/N for calibration. Subsequently, the
program checks if the number of potential calibrant peaks that passed
the S/N check pass the specified minimum number of calibrant peaks prior
to calibrating the chromatogram. This build also addresses a a bug
(Issue #6) by adding an additional check during quantitation (if a file
was not present in the EXCLUSION_fFILES) and by adding a default value
for currNoise in functions.backgroundNoise. This build also introduces a
quality of life feature (Issue #5) by combining an overview chromatogram
and all the individual peaks into a single PDF file.