HappyTools v0.0.2 b171116a
This build introduces automated border/edge determination for each
integration window during automated peak detection. The program allows
the user to specify if the border/edge is determined as the FWHM or if
it is based on a user-specified Sigma value. Two new settings have been
added to the settings window for that purpose.