Hotfix for Bug #12 and basic implementation of Issue #7
This build implements a quick fix for bug #12, by taking the time
difference between ['Data'][-1] and ['Data'][0] as the integration
window. This approach is valid because ['Data'] is based on the
specified edge setting (FWHM or Sigma), which is applied just before
writing the file to disk. Furthermore, this build also implements a
basic restructured settings window (for Issue #7). However, a few
changes still need to be made before that issue can be closed.