HappyTools v0.0.2 b171212a
This build implements updates to the restructured peakDetection function
which fixes bug #17 and most (if not all) of bug #15. Furthermore, the
functions relating to batchPlotting have been edited to fix bug #16. All
these issues will be closed as per this update, please comment in the
relevant issue thread if the problems persist.