HappyTools v0.0.2 b180131a
This build restructures the batchFunctions with the intent of separating
the spaghetti code into clearer/smaller segments. Furthermore, this
build fixes #18 by changing 'Peak':xData[np.argmax(yData)] to
'Peak':newGaussX[np.argmax(newGaussY)] in the functions. Lastly, this
build implements a dynamic guess for the Gaussian parameters that is
based on the actual data rather than a hard coded initial guess.