HappyTools v0.0.2 b180521a
This build fixes bug #21 and implements a new feature (issue #22) that
allows the user to select where the data is located, rather than
expecting the raw data to be present in HappyTools' folder. Furthermore,
this build also implements a function that checks if HappyTools has
read/write access to the folders that it needs for processing.