Happytools v0.0.2 b180725a
This build implements an early version of 'ultraPerformanceCalibration'
which is intended to find the best calibration method for the user
without requiring any prior knowledge. The function currently tries to
fit various polynomials of order 2 to n (number of calibrants-1) and
uses the polynomial function that yields the lowest RMS and ensures that
the calibrated data is monotone.

The 'Interpolation' option of UPC is currently not recommended as it
only ensures monotonicity between the first and last calibrant but not
for the whole measurements' timerange.