HappyTools v0.0.2 b180727a
This build implements an option to use the recently added
ultraPerformanceCalibration, which by default is set to 'False' as it
requires further testing and development. The user is able to enable it,
but it comes with an explicit warning in the tooltip that this is not
ready for production yet.