HappyTools v0.0.2 b180807a
This build continues with the object orientated class refactoring of
HappyTools and has merged the functionality that was previously under
the 'Multi File' menu with the regular 'File' menu of HappyTools, by
allowing several files to be opened and enabling smoothing, baseline
correction and normalization to work on several files at once.

However, the advanced functionality such as peak detection is currently
not functioning in the refactored version. Please use the latest stable
release if that is a functionality that is desired.