HappyTools v0.0.2 b180807b
This builds adds some code to catch some ugly errors, e.g. when the user
tries to smooth a chromatogram while no chromatogram is in memmory.
Furthermore, this build adds a process menu button, from where the user
is able to perform calibration and quantitation on the active
chromatograms (currently linked to place holder functions).