HappyTools v0.0.2 b180814a
This build implements fully functioning `calibration` and `quantitation`
using both the GUI and the 'Batch' menu of HappyTools. Furthermore, this
build adds a new output option to quantitation which reports the area of
the best fitting Gaussian function, which will help getting a more
robust quantitation for (partially) overlapping peaks.