HappyTools v0.0.2 b180815b
This build implements the basic framework of the `Peak` class that
stores all the peak parameters and the functions to determine these
parameters (issue #30). However, the Gaussian functions that are fitted
by this class require the data to be non-negative, which should be not
something the user has to worry about.