HappyTools v0.0.2 b180816a
This build implements a function to the `Peak` class that corrects the
intensities of each peak to the background value that was determined for
the peak, ensuring that no negative intensities are ever encountered.

Furthermore, this build implements a very sloppy progressbar to the main
GUI, that is used during graphical processing of chromatograms. This
should be migrated to the Progressbar class and merged with the `Batch`
progressbars.