Calibrated files will always be '.txt'
HappyTools would retain the extension of the original data when performing calibration during batchprocess. The result was that if a user started with '.awr' or '.csv' files that the quantitation wouldn't occur as that expects a '.txt' file.

This build addresses that by using `Path(location).stem + '.'txt'` instead of `Path(location).name`.