HappyTools v0.1-beta1 b20190305a
This build further implements some refactoring work, one major task is
remaining and that is to tackle the current `Functions` class. The
majority of the code in there should go to their respective classes such
as `Chromatogram` and a new file with BatchProcess related functions
(ideally this would re-use code already present).