HappyTools v0.1-beta1 b190306a
This build further continues with refactoring and migrating code to
their respective classes. This build has deprecated the `Functions`
class and most of it's underlying code has been migrated to the
`Chromatogram` class.

NOTE: This build will generate correct output for GUI-based processing.
However, the batch process functionality is not working.