HappyTools v0.1-beta1 b190310a
This build has re-implemented the batch process functionality, once the
last few items that are still open are finished we will release the
final beta (beta2) as the release candidate.

Tasks that are still left to do:

1. Implement the 'save calibrants' option of the `PeakDetection` class.
2. Move the `ProgressBar` class out of `progress_bar.py` as it should be
in it's own class with a name more fitting it's purpose, such as
`BatchProcessTaskWindow` (name not final).

Stretch goal tasks:

1. Find a way to re-use code of the Chromatogram as to avoid code
duplication in BatchProcess.
2. Implement the option that users do not specify the retention time
window for an analyte, in which case the software should use the value
that will be defined in the `Settings` class.