HappyTools v0.1-beta1 b190311a
This build modifies the `ProgressBar` class as to work with
`self.master` instead of `master.master` which allows it to be more
easily re-used (as long as the right reference is passed upon creation).
Furthermore, this build implements the `BatchProgressProcessWindow`
class, which creates a window which only shows progress bars for steps
that will actually be performed, i.e. if the user only selected
quantitation than no calibration progress bar will be shown.